// src/app/pages/about-me/components/about-me-services/model/service-card-interface.ts
import { ServiceCardIcon } from './service-card-type';

/**
 * Represents a service card.
 */
export interface ServiceCard {
  /**
   * The title of the service card.
   */
  title: string;

  /**
   * The message of the service card.
   */
  msg: string;

  /**
   * The icon of the service card.
   */
  icon: ServiceCardIcon;
}